info on how to gain PUID and GUID with putty.
Nothing too advanced, dont think i should need to note all the steps.
just login with putty with SSH and type in `ID` and you gain the ids.
please correct anything if i did anything wrong.
Felt it was lacking and having it manually set to my own user on Synology helped with folder access on the server.